+++ noodles-bcf/src/record.rs
@@ -1,32 +1,35 @@
 //! BCF record and fields.
 
 mod convert;
+mod filters;
 mod genotypes;
+mod info;
 pub mod value;
 
-pub use self::{genotypes::Genotypes, value::Value};
+pub use self::{filters::Filters, genotypes::Genotypes, info::Info, value::Value};
 
-use std::{
-    io,
-    ops::{Deref, DerefMut},
-};
+use std::io;
 
 use noodles_vcf as vcf;
 
 /// A BCF record.
 ///
 /// A `bcf::Record` wraps a raw byte buffer, and the fields should be considered immutable.
-#[derive(Clone, Default, Eq, PartialEq)]
+#[derive(Clone, Debug, PartialEq)]
 pub struct Record {
-    buf: Vec<u8>,
+    pub(crate) chrom: i32,
+    pub(crate) pos: vcf::record::Position,
+    pub(crate) rlen: i32,
+    pub(crate) qual: Option<vcf::record::QualityScore>,
+    pub(crate) id: vcf::record::Ids,
+    pub(crate) r#ref: vcf::record::ReferenceBases,
+    pub(crate) alt: vcf::record::AlternateBases,
+    filter: Filters,
+    info: Info,
     genotypes: Genotypes,
 }
 
 impl Record {
-    pub(crate) fn resize(&mut self, new_len: usize) {
-        self.buf.resize(new_len, Default::default());
-    }
-
     /// Returns the chromosome ID of the record.
     ///
     /// The chromosome ID is the index of the associated contig in the VCF header.
@@ -34,25 +37,12 @@ impl Record {
     /// # Examples
     ///
     /// ```
-    /// # use std::io;
     /// use noodles_bcf as bcf;
-    ///
-    /// let record = bcf::Record::from(vec![
-    ///     0x08, 0x00, 0x00, 0x00, // CHROM
-    ///     // ...
-    /// ]);
-    ///
-    /// assert_eq!(record.chromosome_id()?, 8);
-    /// # Ok::<(), io::Error>(())
+    /// let record = bcf::Record::default();
+    /// assert_eq!(record.chromosome_id(), 0);
     /// ```
-    pub fn chromosome_id(&self) -> io::Result<i32> {
-        const OFFSET: usize = 0;
-
-        let data = &self.buf[OFFSET..OFFSET + 4];
-
-        data.try_into()
-            .map(i32::from_le_bytes)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+    pub fn chromosome_id(&self) -> i32 {
+        self.chrom
     }
 
     /// Returns the start position of this record.
@@ -63,43 +53,16 @@ impl Record {
     /// # Examples
     ///
     /// ```
-    /// # use std::io;
     /// use noodles_bcf as bcf;
-    /// use noodles_vcf as vcf;
-    ///
-    /// let record = bcf::Record::from(vec![
-    ///     0x08, 0x00, 0x00, 0x00, // CHROM
-    ///     0x0c, 0x00, 0x00, 0x00, // POS
-    ///     // ...
-    /// ]);
-    ///
-    /// assert_eq!(record.position().map(i32::from)?, 13);
-    /// # Ok::<(), io::Error>(())
+    /// let record = bcf::Record::default();
+    /// assert_eq!(i32::from(record.position()), 1);
     /// ```
-    pub fn position(&self) -> io::Result<vcf::record::Position> {
-        use vcf::record::Position;
-
-        const OFFSET: usize = 4;
-
-        let data = &self.buf[OFFSET..OFFSET + 4];
-
-        data.try_into()
-            .map(i32::from_le_bytes)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-            .and_then(|pos| {
-                Position::try_from(pos + 1)
-                    .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
-            })
+    pub fn position(&self) -> vcf::record::Position {
+        self.pos
     }
 
-    fn rlen(&self) -> io::Result<i32> {
-        const OFFSET: usize = 8;
-
-        let data = &self.buf[OFFSET..OFFSET + 4];
-
-        data.try_into()
-            .map(i32::from_le_bytes)
-            .map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+    fn rlen(&self) -> i32 {
+        self.rlen
     }
 
     /// Returns the end position of this record.
@@ -111,26 +74,50 @@ impl Record {
     /// ```
     /// # use std::io;
     /// use noodles_bcf as bcf;
-    /// use noodles_vcf as vcf;
-    ///
-    /// let record = bcf::Record::from(vec![
-    ///     0x08, 0x00, 0x00, 0x00, // CHROM
-    ///     0x0c, 0x00, 0x00, 0x00, // POS
-    ///     0x05, 0x00, 0x00, 0x00, // rlen
-    ///     // ...
-    /// ]);
-    ///
-    /// assert_eq!(record.end().map(i32::from)?, 17);
+    /// let record = bcf::Record::default();
+    /// assert_eq!(record.end().map(i32::from)?, 1);
     /// # Ok::<(), io::Error>(())
     /// ```
     pub fn end(&self) -> io::Result<vcf::record::Position> {
         use vcf::record::Position;
 
-        let start = self.position().map(i32::from)?;
-        let len = self.rlen()?;
+        let start = i32::from(self.position());
+        let len = self.rlen();
         let end = start + len - 1;
 
         Position::try_from(end).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
+    }
+
+    pub(crate) fn quality_score(&self) -> Option<vcf::record::QualityScore> {
+        self.qual
+    }
+
+    pub(crate) fn ids(&self) -> &vcf::record::Ids {
+        &self.id
+    }
+
+    pub(crate) fn reference_bases(&self) -> &vcf::record::ReferenceBases {
+        &self.r#ref
+    }
+
+    pub(crate) fn alternate_bases(&self) -> &vcf::record::AlternateBases {
+        &self.alt
+    }
+
+    pub(crate) fn filters(&self) -> &Filters {
+        &self.filter
+    }
+
+    pub(crate) fn filters_mut(&mut self) -> &mut Filters {
+        &mut self.filter
+    }
+
+    pub(crate) fn info(&self) -> &Info {
+        &self.info
+    }
+
+    pub(crate) fn info_mut(&mut self) -> &mut Info {
+        &mut self.info
     }
 
     pub(crate) fn genotypes(&self) -> &Genotypes {
@@ -142,24 +129,20 @@ impl Record {
     }
 }
 
-impl Deref for Record {
-    type Target = [u8];
+impl Default for Record {
+    fn default() -> Self {
+        use vcf::record::reference_bases::Base;
 
-    fn deref(&self) -> &[u8] {
-        &self.buf
-    }
-}
-
-impl DerefMut for Record {
-    fn deref_mut(&mut self) -> &mut [u8] {
-        &mut self.buf
-    }
-}
-
-impl From<Vec<u8>> for Record {
-    fn from(buf: Vec<u8>) -> Self {
         Self {
-            buf,
+            chrom: 0,
+            pos: vcf::record::Position::try_from(1).unwrap(),
+            rlen: 1,
+            qual: None,
+            id: vcf::record::Ids::default(),
+            r#ref: vcf::record::ReferenceBases::try_from(vec![Base::A]).unwrap(),
+            alt: vcf::record::AlternateBases::default(),
+            filter: Filters::default(),
+            info: Info::default(),
             genotypes: Genotypes::default(),
         }
     }