+++ noodles-bcf/src/record/info.rs
@@ -1,0 +1,27 @@
+#[derive(Clone, Debug, Default, Eq, PartialEq)]
+pub struct Info {
+    buf: Vec<u8>,
+    field_count: usize,
+}
+
+impl Info {
+    pub(crate) fn len(&self) -> usize {
+        self.field_count
+    }
+
+    pub fn set_field_count(&mut self, field_count: usize) {
+        self.field_count = field_count;
+    }
+}
+
+impl AsRef<[u8]> for Info {
+    fn as_ref(&self) -> &[u8] {
+        &self.buf
+    }
+}
+
+impl AsMut<Vec<u8>> for Info {
+    fn as_mut(&mut self) -> &mut Vec<u8> {
+        &mut self.buf
+    }
+}